Address review feedback for retry policy init cast
Agent-Logs-Url: https://github.com/irgordon/kernul/sessions/de2e1a96-0dd4-4246-ae56-f6bea82b8ff4

+++ core/session.c
@@ -119,7 +119,7 @@ struct session *session_create(struct process *leader)
                      0U,
                      __ATOMIC_RELEASE);
     __atomic_store_n(&session_slot.retry_policy_state,
-                     (u32)SESSION_RETRY_UNSET,
+                     SESSION_RETRY_UNSET,
                      __ATOMIC_RELEASE);
     __atomic_store_n(&session_slot.ready_published, 0U, __ATOMIC_RELAXED);
     __atomic_store_n(&session_slot.terminal_cause,
